refactor: Update dependencies and remove unused code in NewProduct.tsx, Search.tsx, SingleCategory.tsx, Products.tsx, and usePostProducts.ts
