+++ Ecomerce/src/main.tsx
@@ -3,8 +3,8 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import Home from "./pages/Home.tsx";
-import NewProduct from "./pages/NewProduct.tsx";
 import DetailProduct from "./pages/DetailProduct.tsx";
+import NewProductPage from "./pages/NewProductPage.tsx";
 
 const router = createBrowserRouter([
   {
@@ -12,8 +12,12 @@ const router = createBrowserRouter([
     element: <Home></Home>,
   },
   {
+    path: "/home",
+    element: <Home></Home>,
+  },
+  {
     path: "/new",
-    element: <NewProduct />,
+    element: <NewProductPage />,
   },
   {
     path: "/product/:categoryId/:productId",
