+++ Ecomerce/src/pages/Home.tsx
@@ -1,11 +1,14 @@
 import Navbar from "../components/Navbar";
 import Products from "../components/Products";
+import ProductsProvider from "../context/ProductsProvider";
 
 export default function Home() {
   return (
     <>
-      <Navbar></Navbar>
-      <Products></Products>
+      <ProductsProvider>
+        <Navbar></Navbar>
+        <Products></Products>
+      </ProductsProvider>
     </>
   );
 }
